Feat/order

See merge request cloud_track/class_02/web_project2/team04/singleshop!19

+++ src/main/java/com/elice/team4/singleShop/order/dto/OrderDto.java
@@ -1,5 +1,6 @@
 package com.elice.team4.singleShop.order.dto;
 
+import jakarta.validation.constraints.NotNull;
 import lombok.Getter;
 import lombok.Setter;
 
@@ -7,7 +8,8 @@ import lombok.Setter;
 @Setter
 public class OrderDto {
 
-    private int productId;
+    //@NotNull(message = "상품 아이디는 필수값입니다.")
+    private Long productId;
 
     private int count;
 }
